+++ .eslintrc
@@ -34,7 +34,7 @@
     "class-methods-use-this": 0,
     "no-unused-expressions": ["error", { "allowTaggedTemplates": true }],
     "no-underscore-dangle": ["error", {
-      "allow": ["__REDUX_DEVTOOLS_EXTENSION__"]
+      "allow": ["__REDUX_DEVTOOLS_EXTENSION_COMPOSE__"]
     }],
     "react/no-unused-prop-types": 0,
     "consistent-return": 0,
