+++ containers/AuthFields/store.js
@@ -1,6 +1,6 @@
 import persist from '../../libraries/persist';
 
-// Actions Naming
+// Constants
 export const AUTH_SIGNIN = 'AUTH_SIGNIN';
 export const AUTH_SIGNOUT = 'AUTH_SIGNOUT';
 export const AUTH_SERVERERROR = 'AUTH_SERVERERROR';
@@ -24,6 +24,7 @@ const reducer = (state = initialState, action) => {
       };
     case AUTH_SIGNOUT:
       return { ...state, authenticated: false, token: null, error: null };
+    // TODO: not used? can remove?
     case AUTH_SERVERERROR:
       return { ...state, authenticated: false, error: action.error };
     default:
@@ -31,23 +32,23 @@ const reducer = (state = initialState, action) => {
   }
 };
 
-// Actions
-const actions = {};
+// Action creators
+const actionCreators = {};
 
-actions.signIn = token => ({ type: AUTH_SIGNIN, token });
-actions.signOut = () => ({ type: AUTH_SIGNOUT });
+actionCreators.signIn = token => ({ type: AUTH_SIGNIN, token });
+actionCreators.signOut = () => ({ type: AUTH_SIGNOUT });
 
 // Discpatchers
 const dispatchers = {};
 
 dispatchers.signIn = token => {
   persist.willSetAccessToken(token);
-  return actions.signIn(token);
+  return actionCreators.signIn(token);
 };
 
 dispatchers.signOut = () => {
   persist.willRemoveAccessToken();
-  return actions.signOut();
+  return actionCreators.signOut();
 };
 
-export { actions, reducer, dispatchers };
+export { actionCreators, reducer, dispatchers };
