+++ containers/AuthFields/styles.js
@@ -1,0 +1,20 @@
+import styled from 'styled-components';
+
+export const Main = styled.div`
+  border-bottom: 1px solid #ececec;
+  padding-bottom: 20px;
+  margin-bottom: 20px;
+
+  > h1 {
+    font-size: 20px;
+  }
+
+  > input {
+    display: block;
+    margin-bottom: 10px;
+  }
+`;
+
+export const SubmitButton = styled.button`
+  opacity: ${({ touched }) => (touched ? 1 : 0.5)};
+`;
