+++ containers/CreatePost/index.js
@@ -1,0 +1,4 @@
+import Feature from './feature';
+import withApollo from './apollo';
+
+export default withApollo(Feature);
