+++ containers/CreatePost/styles.js
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+// eslint-disable-next-line import/prefer-default-export
+export const Form = styled.form`
+  border-bottom: 1px solid #ececec;
+  padding-bottom: 20px;
+  margin-bottom: 20px;
+  > h1 {
+    font-size: 20px;
+  }
+  > input {
+    display: block;
+    margin-bottom: 10px;
+  }
+`;
