+++ containers/PostInfo/index.js
@@ -1,0 +1,4 @@
+import withApollo from './apollo';
+import Feature from './feature';
+
+export default withApollo(Feature);
