+++ containers/PostInfo/styles.js
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+// eslint-disable-next-line import/prefer-default-export
+export const Section = styled.section`
+  padding-bottom: 20px;
+
+  > h1 {
+    margin-top: 0;
+  }
+
+  > p {
+    font-size: 17px;
+  }
+`;
