+++ libraries/reducer.js
@@ -1,5 +1,5 @@
 import { combineReducers } from 'redux';
-import { reducer as authReducer } from '../components/AuthFields/index.data';
+import { reducer as authReducer } from '../containers/AuthFields/store';
 
 export default function getReducer(client) {
   return combineReducers({
