+++ package.json
@@ -8,7 +8,8 @@
     "dev:next": "npm run lint && next",
     "start": "NODE_ENV=production node server.js",
     "start:next": "next start",
-    "lint": "eslint --fix --ext .gql --ext .graphql --ext .js pages components libraries server helper_scripts",
+    "lint": "eslint --fix --ext .gql --ext .graphql --ext .js pages components containers layouts libraries server helper_scripts",
+    "lint:watch": "esw -w --fix --ext .gql --ext .graphql --ext .js pages components containers layouts libraries server helper_scripts",
     "analyze": "ANALYZE_BUILD=true npm run build",
     "setup": "node ./helper_scripts/CL_commands/setup.js && npm run setup:after",
     "setup:after": "npm run build",
@@ -93,6 +94,7 @@
     "eslint-plugin-jsx-a11y": "5.1.1",
     "eslint-plugin-prettier": "2.1.2",
     "eslint-plugin-react": "7.1.0",
+    "eslint-watch": "^3.1.2",
     "figlet": "1.2.0",
     "graphql-cli": "1.0.0-beta.4",
     "graphql-cli-voyager": "0.1.2",
