+++ pages/index.js
@@ -1,6 +1,6 @@
-import PostList from '../components/PostList';
+import PostList from '../containers/PostList';
 import withData from '../libraries/withData';
-import DefaultCon from '../containers/Default';
+import DefaultCon from '../layouts/Default';
 
 export default withData(props =>
   <DefaultCon {...props}>
