+++ pages/signin.js
@@ -1,6 +1,6 @@
+import SignInForm from '../containers/SignInForm';
 import withData from '../libraries/withData';
-import SignInForm from '../components/SignInForm';
-import DefaultCon from '../containers/Default';
+import DefaultCon from '../layouts/Default';
 
 export default withData(props =>
   <DefaultCon title="Sign In" {...props}>
